Minutely.precipitation rettet fra int til double

+++ app/src/main/java/dk/simonpeter/weatherforecastapp/openweathermap/onecall/Minutely.kt
@@ -2,5 +2,5 @@ package dk.simonnpeter.weatherforecastapp.openweathermap.onecall
 
 data class Minutely(
     val dt: Int,
-    val precipitation: Int
+    val precipitation: Double
 )
